Refactor chat system to use session-based routing and messaging
