+++ package.json
@@ -77,8 +77,8 @@
     "date-fns": "^4.1.0",
     "dockerode": "^4.0.6",
     "dotenv": "^16.5.0",
-    "drizzle-kit": "^0.31.0",
-    "drizzle-orm": "^0.43.1",
+    "drizzle-kit": "^0.31.4",
+    "drizzle-orm": "^0.44.4",
     "embla-carousel-react": "^8.6.0",
     "framer-motion": "^12.9.2",
     "get-port": "^7.1.0",
